chore: rename .mjs to .js across ESM packages
With every package now `type: "module"`, the .mjs extension is redundant
— `.js` already means ESM. Standardize on .js for all source/config
files in inspector, package eslint configs, and the root prettier config.

https: //claude.ai/code/session_01DxhZ1QN3nUgrHvoXvKsPKV
