+++ apps/inspector/manifest.js
