+++ packages/core/eslint.config.js
