+++ packages/model/eslint.config.js
